+++ source/extensions/isaacsim.examples.interactive/docs/CHANGELOG.md
@@ -1,12 +1,17 @@
 # Changelog
-## [4.9.3] - 2025-11-27
-### Changed
-- Add missing docstrings
+## [4.11.0] - 2025-12-02
+### Removed
+- base_sample_experimental.py
+- base_sample_extension_experimental.py
 
 ## [4.10.0] - 2025-12-01
 ### Changed
 - Updated inference examples to use GPU physics and the new experimental APIs
 - Moved policy based examples to isaacsim.robot.policy.examples
+
+## [4.9.3] - 2025-11-27
+### Changed
+- Add missing docstrings
 
 ## [4.9.2] - 2025-11-24
 ### Changed